Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-gainnode-interface/gain-expected.txt

CONSOLE MESSAGE: line 85: TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
                context, {numberOfInputs: context.destination.channelCount})')

Harness Error (FAIL), message = TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
                context, {numberOfInputs: context.destination.channelCount})')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [create context] Create context for test 
PASS < [create context] All assertions passed. (total 0 assertions) 
PASS > [test] GainNode functionality 
